+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - #531 : Dynamic interface mixin 
   - #534 : Generic typecast handler 
   - #535 : Setting several properties in a row 
+  - #530 : Supporting unary operators : ! and -
 - ELC
   - a warning shown when an object cannot be converted into another one
   - a warning shown when an object does not handle a message (for strong-typed ones)